build: use rollup's `preserveModules` option
This makes it possible for a package user to dig into the dist folder for a certain export (like `parseTwitchMessage`) if they cant import it otherwise (not exported or non-node environments).

Also allows possibility of further expanding of the `exports` fields in package.json, and makes declaration files match one for one.